+++ spec/rubocop/cop/style/self_assignment_spec.rb
@@ -1,0 +1,25 @@
+# encoding: utf-8
+
+require 'spec_helper'
+
+describe Rubocop::Cop::Style::SelfAssignment do
+  subject(:cop) { described_class.new }
+
+  described_class::OPS.product(['x', '@x', '@@x']).each do |op, var|
+    it "registers an offense for non-shorthand assignment #{op} and #{var}" do
+      inspect_source(cop,
+                     ["#{var} = #{var} #{op} y"])
+      expect(cop.offenses.size).to eq(1)
+      expect(cop.messages)
+        .to eq(["Use self-assignment shorthand #{op}=."])
+    end
+
+    it "accepts shorthand assignment for #{op} and #{var}" do
+      inspect_source(cop,
+                     ["#{var} = #{var} #{op} y"])
+      expect(cop.offenses.size).to eq(1)
+      expect(cop.messages)
+        .to eq(["Use self-assignment shorthand #{op}=."])
+    end
+  end
+end