+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * New cop `DeprecatedClassMethods` checks for deprecated class methods. ([@bbatsov][])
 * New cop `StringConversionInInterpolation` checks for redundant `Object#to_s` in string interpolation. ([@bbatsov][])
 * New cop `LiteralInInterpolation` checks for interpolated string literals. ([@bbatsov][])
+* New cop `SelfAssignment` checks for places where the self-assignment shorthand should have been used. ([@bbatsov][])
 * New Rails cop `ActionFilter` enforces the use of `_filter` or `_action` action filter methods. ([@bbatsov][])
 * New Rails cop `ScopeArgs` makes sure you invoke the `scope` method properly. ([@bbatsov][])
 * [#743](https://github.com/bbatsov/rubocop/issues/743): `SingleLineMethods` cop does auto-correction. ([@jonas054][])